feat: implement search & sort based price

+++ Simple_Bookstore/src/controllers/book.php
@@ -34,13 +34,25 @@ class Book extends Validation
      {
           try {
                $data = $_POST;
-               $this->validate($data, [
-                    'title' => ['required', 'string'],
-                    'author' => ['required', 'string'],
-                    'price' => ['required', 'number']
-               ]);
+               $validationRules = [
+                    'name' => 'required|string|min:3',
+                    'brand' => 'required|string|min:3',
+                    'price' => 'required|number|min:1'
+               ];
+
+               foreach ($validationRules as $key => $rule) {
+                    $this->addRule($key, $rule);
+               }
+
+               $error = $this->validate($data);
+               if (!empty($error)) {
+                    $this->controller->render('index', [
+                         'books' => $this->bookModel->getAll(),
+                         'error' => $error
+                    ]);
+               }
                $this->bookModel->create($data);
-               $this->controller->redirect('book');
+               $this->controller->redirect('/api/v1/book', $error);
           } catch (PDOException $e) {
                $this->controller->render('error', ['error' => $e->getMessage()]);
           } catch (Exception $e) {
@@ -50,12 +62,26 @@ class Book extends Validation
      public function update($id)
      {
           try {
+               $book = $this->bookModel->getById($id);
+               if (empty($book)) throw new Exception('Book not found');
                $data = $_POST;
-               $this->validate($data, [
-                    'title' => ['required', 'string'],
-                    'author' => ['required', 'string'],
-                    'price' => ['required', 'number']
-               ]);
+               $validationRules = [
+                    'name' => 'required|string|min:3',
+                    'brand' => 'required|string|min:3',
+                    'price' => 'required|number|min:1'
+               ];
+
+               foreach ($validationRules as $key => $rule) {
+                    $this->addRule($key, $rule);
+               }
+
+               $error = $this->validate($data);
+               if (!empty($error)) {
+                    $this->controller->render('update', [
+                         'book' => $this->bookModel->getById($id),
+                         'error' => $error
+                    ]);
+               }
                $this->bookModel->update($id, $data);
                $this->controller->redirect('/api/v1/book');
           } catch (PDOException $e) {
@@ -67,9 +93,39 @@ class Book extends Validation
      public function delete($id)
      {
           try {
+               $book = $this->bookModel->getById($id);
+               if (empty($book)) throw new Exception('Book not found');
                $this->bookModel->delete($id);
                $this->controller->redirect('/api/v1/book');
           } catch (PDOException $e) {
+               $this->controller->render('error', ['error' => $e->getMessage()]);
+          }
+     }
+     public function deleteAll()
+     {
+          try {
+               $this->bookModel->deleteAll();
+               $this->controller->redirect('/api/v1/book');
+          } catch (PDOException $e) {
+               $this->controller->render('error', ['error' => $e->getMessage()]);
+          }
+     }
+     public function searchAndFilterByType()
+     {
+          try {
+               $postData = [
+                    'simple-search' => $_GET['simple-search'] ?? '',
+                    'price' => $_GET['price'] ?? 'low'
+               ];
+               $books = $this->bookModel->searchAndFilter($postData);
+               $this->controller->render('index', [
+                    'books' => $books,
+                    'searchTerm' => $postData['simple-search'],
+                    'priceOrder' => $postData['price']
+               ]);
+          } catch (PDOException $e) {
+               $this->controller->render('error', ['error' => $e->getMessage()]);
+          } catch (Exception $e) {
                $this->controller->render('error', ['error' => $e->getMessage()]);
           }
      }
